Implement the first draft version of the sonar-phab communication
Summary: Quick and dirty implementation to get the sonar-phab communication working, based on the previous gradle plugin code. Ran into some issues with gradle that should not be a problem with an exe, but I haven't resolved them yet. Plan is to run this as a proof of concept first, and then  clean it up later for a version 1.0 (yes, famous last words).

Test Plan: N/A

Reviewers: sergei.prodanov

Reviewed By: sergei.prodanov

JIRA Issues: TD-232

Differential Revision: https://phabricator.elhub.cloud/D1140